% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_activities.R
\name{build_activities}
\alias{build_activities}
\title{Build an activities dataframe from the NHTS Trips dataset}
\usage{
build_activities(trips)
}
\arguments{
\item{trips}{A `tibble` containing records from the `nhts_trips` dataset. Must
contain (at a minumum) the following columns:
- `houseid`
- `personid`
- `strttime`
- `endtime`
- `whyfrom`
- `whyto`}
}
\value{
A `tibble` with one row per activity for each person in the `trips`
  input dataset.
}
\description{
Build an activities dataframe from the NHTS Trips dataset
}
\details{
The NHTS data does not have a start time for the travel day. This
  function assumes that all days will begin at 4:00 AM, though this may introduce
  illogical activity patterns in some situations.
}
\examples{

}
\author{
Nate Lant
}
