% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tours.R
\name{build_tours}
\alias{build_tours}
\title{Determine the tour, tour type, and daily activity pattern from a list of activities}
\usage{
build_tours(activities)
}
\arguments{
\item{activities}{A `tibble` constructed from `build_activities`}
}
\value{
A `tibble` containing the details of the tours and DAP for the
  supplied activities.
}
\description{
Determine the tour, tour type, and daily activity pattern from a list of activities
}
