% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhts2017.R
\docType{package}
\name{nhts2017}
\alias{nhts2017}
\title{nhts2017: A package for distributing the 2017 National Household Travel Survey}
\description{
The datasets in this package are tidy tibbles with proper factors and definitions
encoded.
}
