% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhts_households}
\alias{nhts_households}
\title{Public Household Records}
\format{
A `tibble` with 129,696 rows and 58 variables:
  - `houseid`: Household Identifier
  - `travday`: Travel day - day of week
  - `sampstrat`: Primary Sampling Stratum Assignment
  - `homeown`: Home Ownership
  - `hhsize`: Count of household members
  - `hhvehcnt`: Count of household vehicles
  - `hhfaminc`: Household income
  - `pc`: Frequency of Desktop or Laptop Computer Use to Access the Internet
  - `sphone`: Frequency of Smartphone Use to Access the Internet
  - `tab`: Frequency of Tablet Use to Access the Internet
  - `walk`: Frequency of Walking for Travel
  - `bike`: Frequency of Bicycle Use for Travel
  - `car`: Frequency of Personal Vehicle Use for Travel
  - `taxi`: Frequency of Taxi Service or Rideshare Use for Travel
  - `bus`: Frequency of Bus Use for Travel
  - `train`: Frequency of Train Use for Travel
  - `para`: Frequency of Paratransit Use for Travel
  - `price`: Price of Gasoline Affects Travel
  - `place`: Travel is a Financial Burden
  - `walk2save`: Walk to Reduce Financial Burden of Travel
  - `bike2save`: Bicycle to Reduce Financial Burden of Travel
  - `ptrans`: Public Transportation to Reduce Financial Burden of Travel
  - `hhrelatd`: At least two household persons are related
  - `drvrcnt`: Number of drivers in household
  - `cnttdhh`: Count of household trips on travel day
  - `hhstate`: Household state
  - `hhstfips`: State FIPS for household address
  - `numadlt`: Count of adult household members at least 18 years old
  - `youngchild`: Count of persons with an age between 0 and 4 in household
  - `wrkcount`: Number of workers in household
  - `tdaydate`: Date of travel day (YYYYMM)
  - `hhresp`: Person identifier of household respondent
  - `lif_cyc`: Life Cycle classification for the household, derived by attributes pertaining to age, relationship, and work status.
  - `msacat`: Metropolitan Statistical Area (MSA) category for the household's home address, based on household's home geocode and TIGER/Line Shapefiles.
  - `msasize`: Population size category of the Metropolitan Statistical Area (MSA), from the 2010-2014 five-year American Community Survey (ACS) API.
  - `rail`: MSA heavy rail status for household
  - `urban`: Household's urban area classification, based on home address and 2014 TIGER/Line Shapefile
  - `urbansize`: Urban area size where home address is located
  - `urbrur`: Household in urban/rural area
  - `scresp`: Person identifier of mail screener respondent, always 1 to roster self first
  - `census_d`: 2010 Census division classification for the respondent's home address
  - `census_r`: Census region classification for home address
  - `cdivmsar`: Grouping of household by combination of census division, MSA status, and presence of a subway system when population greater than 1 million
  - `hh_race`: Race of household respondent
  - `hh_hisp`: Hispanic status of household respondent
  - `hh_cbsa`: Core Based Statistical Area (CBSA) FIPS code for the respondent's home address
  - `resp_cnt`: Count of responding persons per household
  - `webuse17`: Frequency of internet use
  - `smplsrce`: Sample where the case originated
  - `wthhfin`: Final HH weight
  - `hbhur`: Urban / Rural indicator -  Block group
  - `hthtnrnt`: Category of the percent of renter-occupied housing in the census tract of the household's home location.
  - `htppopdn`: Category of population density (persons per square mile) in the census tract of the household's home location.
  - `htresdn`: Category of housing units per square mile in the census tract of the household's home location.
  - `hteempdn`: Category of workers per square mile in the census tract of the household's home location.
  - `hbhtnrnt`: Category of the percent of renter-occupied housing in the census block group of the household's home location.
  - `hbppopdn`: Category of population density (persons per square mile) in the census block group of the household's home location.
  - `hbresdn`: Category of housing units per square mile in the census block group of the household's home location.
}
\usage{
nhts_households
}
\description{
A dataset containing households records from the 2017 National household
travel survey.
}
\keyword{datasets}
