% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhts_persons}
\alias{nhts_persons}
\title{Public Person Records}
\format{
A `tibble` with 264,234 rows and 87 variables:
  - `houseid`: Household Identifier
  - `personid`: Person Identifier
  - `r_age`: Age
  - `educ`: Educational Attainment
  - `r_hisp`: Hispanic or Latino Origin
  - `r_relat`: Relationship
  - `r_sex`: Gender
  - `r_race`: Race
  - `prmact`: Primary Activity in Previous Week
  - `payprof`: Work for Pay in Previous Week
  - `gt1jblwk`: More than One Job
  - `wrk_home`: Work from Home
  - `wkftpt`: Full-Time or Part-Time Worker
  - `wrktrans`: Mode to Work
  - `lsttrday17`: Last trip before travel day
  - `occat`: Job Category
  - `schtyp`: Student Status
  - `nwalktrp`: Count of Walk Trips
  - `walk4ex`: Count of Walk Trips for Exercise
  - `nbiketrp`: Count of Bike Trips
  - `bike4ex`: Count of Bike Trips for Exercise
  - `bikeshare`: Count of Bike Share Program Usage
  - `ptused`: Count of Public Transit Usage
  - `mcused`: Count of Motorcycle or Moped Trips
  - `carshare`: Count of Car Share Program Usage
  - `rideshare`: Count of Rideshare App Usage
  - `carrode`: Count of People in Vehicle to Work
  - `timetowk`: Trip Time to Work in Minutes
  - `nocong`: Trip Time in Minutes to Work without Traffic
  - `pubtime`: Minutes Spent Transferring to Work
  - `wrktime`: Arrival Time at Work
  - `wkrmhm`: Option of Working from Home
  - `flextime`: Flex Time
  - `wkfmhmxx`: Count of Days Worked From Home in Last Month
  - `schtrn1`: Mode to School
  - `schtrn2`: Mode from School
  - `deliver`: Count of Times Purchased Online for Delivery in Last 30 Days
  - `medcond`: Medical Condition
  - `medcond6`: Medical Condition, How Long
  - `health`: Opinion of Health
  - `phyact`: Level of Physical Activity
  - `vpact`: Count of Times of Vigorous Physical Activity in Past Week
  - `lpact`: Count of Times of Light or Moderate Physical Activity in Past Week
  - `borninus`: Born in United States
  - `yrtous`: Year of Arrival in United States
  - `yearmile`: Miles Personally Driven in all Vehicles
  - `proxy`: Trip info from respondent or proxy
  - `whoproxy`: Household Person Identifier Responsible for Trip Reporting
  - `usepubtr`: Public Transit Usage on Travel Date, derived
  - `sameplc`: Reason for No Trips
  - `w_none`: Medical Device Used: None
  - `w_cane`: Medical Device Used: Cane
  - `w_wlkr`: Medical Device Used: Walker
  - `w_whcane`: Medical Device Used: White Cane
  - `w_dog`: Medical Device Used: Dog Assistance
  - `w_crutch`: Medical Device Used: Crutches
  - `w_scootr`: Medical Device Used: Motorized Scooter
  - `w_chair`: Medical Device Used: Wheelchair
  - `w_mtrchr`: Medical Device Used: Motorized Wheelchair
  - `worker`: Worker status
  - `diary`: Travel Diary completion status
  - `outcntry`: Respondent was out of the country on travel day.
  - `frsthm17`: Travel day began at home location
  - `cnttdtr`: Count of person trips on travel day
  - `gcdwork`: Minimum geodesic (Great Circle) distance between home location and work location in meters, using WGS84 coordinate system
  - `wkstfips`: The state FIPS code for the respondent's geocoded work address. The state FIPS codes were identified using United States Census Bureau 2016 TIGER/Line Shapefiles.
  - `driver`: Driver status, derived
  - `outoftwn`: Away from home for entire travel day
  - `disttowk17`: Road network distance, in miles, between respondent's home location and work location, sourced using Google Distance Matrix API (https://maps.googleapis.com/maps/api/distancematrix/)
  - `disttosc17`: Road network distance, in miles, between respondent's home location and school location, sourced using Google Distance Matrix API (https://maps.googleapis.com/maps/api/distancematrix/)
  - `r_age_imp`: Age (imputed)
  - `r_sex_imp`: Gender (imputed)
  - `alt_16`: Alternative Mode of Transportation: Public Transportation or Taxi
  - `alt_23`: Alternative Mode of Transportation: Passenger to Friend/Family Member or Rental Car
  - `alt_45`: Alternative Mode of Transportation: Bicycle or Walk
  - `walk_def`: Reason for Not Walking More: Infrastructure
  - `walk_gkq`: Reason for Not Walking More: Safety
  - `bike_dfr`: Reason for Not Biking More: Infrastructure
  - `bike_gkp`: Reason for Not Biking More: Safety
  - `condtrav`: Medical Condition Results in Reduced Day-to-Day Travel
  - `condride`: Medical Condition Results in Asking Others for Rides
  - `condnigh`: Medical Condition Results in Limiting Driving to Daytime
  - `condrive`: Medical Condition Results in Giving up Driving
  - `condpub`: Medical Condition Results in Using Bus or Subway Less Frequently
  - `condspec`: Medical Condition Results in Using Special Transportation Services
  - `condtax`: Medical Condition Results in Using a Reduced Fare Taxi
  - `wtperfin`: Final Person weight
}
\usage{
nhts_persons
}
\description{
A dataset containing person records for
}
\keyword{datasets}
