% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhts_trips}
\alias{nhts_trips}
\title{Public Trip Records}
\format{A `tibble` with N rows and K variables:
  - `houseid`: Household Identifier
  - `personid`: Person Identifier
  - `trvlcmin`: Trip Duration in Minutes
  - `trpmiles`: Trip distance in miles, derived from route geometry returned by Google Maps API, or from reported loop-trip distance
  - `trptrans`: Trip Mode, derived
  - `trpaccmp`: Count of People on Trip
  - `trphhacc`: Count of Household Members on Trip
  - `vehid`: Household Vehicle Identifier Used on Trip
  - `trwaittm`: Transit wait time in minutes
  - `numtrans`: Count of Transfers
  - `tracctm`: Trip Time to Transit Station in Minutes
  - `drop_prk`: Parked or Dropped Off at Station
  - `tregrtm`: Time to Destination from Transit in Minutes
  - `whodrove`: Person Identifier Who Drove on Trip
  - `whyfrom`: Trip Origin Purpose
  - `loop_trip`: Trip origin and destination at identical location
  - `trphhveh`: Household Vehicle Used on Trip
  - `hhmemdrv`: Household member drove on trip
  - `hh_ontd`: Number of household members on trip including respondent
  - `nonhhcnt`: Number of non-household members on trip
  - `numontrp`: Number of people on trip including respondent
  - `psgr_flg`: Respondent was passenger on trip
  - `pubtrans`: Public transportation used on trip
  - `trippurp`: Generalized purpose of trip, home-based and non-home based
  - `dweltime`: Time at destination
  - `tdwknd`: Weekend trip
  - `vmt_mile`: Trip distance in miles for personally driven vehicle trips, derived from route geometry returned by Google Maps API
  - `drvr_flg`: Respondent drove on trip
  - `whytrp1s`: Trip purpose summary
  - `ontd_p1`: Household Person Identifier on Trip: 1
  - `ontd_p2`: Household Person Identifier on Trip: 2
  - `ontd_p3`: Household Person Identifier on Trip: 3
  - `ontd_p4`: Household Person Identifier on Trip: 4
  - `ontd_p5`: Household Person Identifier on Trip: 5
  - `ontd_p6`: Household Person Identifier on Trip: 6
  - `ontd_p7`: Household Person Identifier on Trip: 7
  - `ontd_p8`: Household Person Identifier on Trip: 8
  - `ontd_p9`: Household Person Identifier on Trip: 9
  - `ontd_p10`: Household Person Identifier on Trip: 10
  - `ontd_p11`: Household Person Identifier on Trip: 11
  - `ontd_p12`: Household Person Identifier on Trip: 12
  - `ontd_p13`: Household Person Identifier on Trip: 13
  - `tdcaseid`: Unique identifier for every trip record in the file
  - `tracc_wlk`: Walk as mode used to get to public transit?
  - `tracc_pov`: POV as mode used to get to public transit?
  - `tracc_bus`: Bus as mode used to get to public transit?
  - `tracc_crl`: Rail as mode used to get to public transit?
  - `tracc_sub`: Subway as mode used to get to public transit?
  - `tracc_oth`: Other mode used to get to public transit?
  - `tregr_wlk`: Walk as mode used to get from public transit?
  - `tregr_pov`: POV as mode used to get from public transit?
  - `tregr_bus`: Bus as mode used to get from public transit?
  - `tregr_crl`: Rail as mode used to get from public transit?
  - `tregr_sub`: Subway as mode used to get from public transit?
  - `tregr_oth`: Other mode used to get from public transit?
  - `whyto`: Trip Destination Purpose
  - `gasprice`: Price of gasoline, in cents, on respondent's travel day
  - `wttrdfin`: Final trip weight
  - `whytrp90`: Travel day trip purpose consistent with 1990 NPTS design.}
\usage{
nhts_trips
}
\description{
A dataset containing trip records for
}
\keyword{datasets}
