% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhts_vehicles}
\alias{nhts_vehicles}
\title{Public Vehicle Records}
\format{
A `tibble` with 256,115 rows and 20 variables:
  - `houseid` :Household Identifier
  - `vehid` :Vehicle Identifier
  - `vehyear` :Vehicle Year
  - `vehage` :Age of vehicle, based on model year
  - `make` :Vehicle Make
  - `model` :Vehicle Model
  - `fueltype` :Fuel Type
  - `vehtype` :Vehicle Type
  - `whomain` :Vehicle Main Driver
  - `od_read` :Odometer Reading
  - `hfuel` :Type of Hybrid Vehicle
  - `vehowned` :Owned Vehicle Longer than a Year
  - `vehownmo` :Months of Vehicle Ownership
  - `annmiles` :Self-reported annualized mile estimate
  - `hybrid` :Hybrid vehicle
  - `personid` :Person ID of Main Driver of this vehicle
  - `bestmile` :Best estimate of annual miles
  - `best_flg` :How BESTMILE was computed
  - `best_edt` :Flag any edits/adjustments to BESTMILE
  - `best_out` :Flag identifying BESTMILE outlier values
}
\usage{
nhts_vehicles
}
\description{
A dataset containing vehicle records from the 2017 National Household
Travel Survey
}
\keyword{datasets}
